WIP: network, bind mech, tests: export some funcs to a test helper pkg
